readability.c
Current student of Harvard University's CS50 Introduction to Computer Science.  In week 3 of the class we were to create a program, where we could determine the grade "reading level."  Also known as the Coleman-Liau Index, this calculation takes a paragraph: counts the number of letters, words, and sentences to determine the reading level based on this formula:

index = 0.0588 * L - 0.296 * S - 15.8

So our software had to use this formula, accept a paragraph from the user, and return a "grade level rating" based on the passage supplied from the user.